Save Manager
Added a basic save framework that uses a full string passed to it. Did not do anything with actually building save file content, just the file handling functions.

Includes async upload routines for future support of uploading saves to a URL.

Included an experimental background animation on the Save/Load window and added in the Sidebar panel for the menu buttons.

Made a pre-emptive couple fixes to ItemManager